old-tests/submission/Infraware/Forms/contents/Forms/meter_min.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>meter_min</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
		 <p>meter: <meter max=100 min=0 id='meter_id'></meter></p>
	</form>

	<script>
		
		var meter = document.getElementById("meter_id");
		meter.value = 50;
		var min = meter.min;
		meter.min = 20;

		if (typeof(meter.min) == "number") {
			if (min == 0 && meter.min == 20) {
				test(function() {
					assert_true(true, "min attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("min attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("min attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
